Split projects between - reverted old app

+++ src/backend-express/rest-api/employee/banUser.js
@@ -5,13 +5,19 @@ const banUser = async (req, res, usersCollection, ObjectId) => {
     const isValidLogin = await verifyAuth(req, res);
 
     if (isValidLogin === true) {
+      const role = req.cookies.roleType;
 
-      const deletedUser = await usersCollection.deleteOne({ _id: new ObjectId(req.params.id) });
+      if (role === "employee") {
+        const deletedUser = await usersCollection.deleteOne({ _id: new ObjectId(req.params.id) });
 
-      if (deletedUser.deletedCount === 1) {
-        res.json({ status: 'success' });
+        if (deletedUser.deletedCount === 1) {
+          res.json({ status: 'success' });
+        } else {
+          res.status(404).json({ error: "Nie znaleziono użytkownika o podanym id" });
+        }
+
       } else {
-        res.status(404).json({ error: "Nie znaleziono użytkownika o podanym id" });
+        res.status(401).json({ error: "Nie masz wymaganych uprawnień" });
       }
     }
   } catch (err) {
